Make Message an interface
This is pursuant to supporting more complex message-creation behavior
for reasons of efficiency.  The old concrete Message implementation
has been embellished to create "hybridMessage," which is for
supporting large messages without full in-memory buffering, although
for now both it and the .Next() method of MessageStreams have not yet
been taught how to do that.

Signed-off-by: Dan Farina <drfarina@acm.org>