Tweak Message internals for buffered use cases
Message now has slightly different mechanics to try to support
fully-buffered use cases better, such as in the case of deep message
inspection.  Right now, if one parses a message the stream it came
from is exhausted, but there are times where it is useful to parse a
message and then subsequently relay it.  It is also possible to know
a-priori that the message is not too large based on its header.

Signed-off-by: Dan Farina <drfarina@acm.org>