Remove IsStartup operator
It has a DOS exploit, because it doesn't check message size before
.Force().  It also can be wrong, because it doesn't actually verify
the structure of the startup packet, which ReadStartupMessage does
properly.

Finally, this operator would only be useful to the hypothetical person
who wanted to know something is kind-of-like a startup packet (right
size, right header) without parsing it.  Until then, it's gone.

Signed-off-by: Daniel Farina <daniel@heroku.com>