Handle the last read of data before error
Before this, an error would cause the last buffer's worth of read
messages to be lost.  Now the implementation should properly give you
as many bytes as it was able to get before the error, squeezing out
the final messages before transitioning to an error state.

This could have been caught by a test that does message
fragmentation-fuzzing, and require more time than I have at this
moment to mock out a test, so there is no regression test for this
case at this time.

Signed-off-by: Dan Farina <drfarina@acm.org>