Implement and switch to serialization-based regression testing
Implements a tiny framework for freezing the results of operators over
data, in this case the resultant AST from parsing the dogs command
language.

A modified "fmt" package from Go is used to do this, to overcome
problems with non-deterministic output order of maps.

Output of the astRegress function ends up in the 'results' directory,
and then compared to their equivalents in the 'expected' directory.

Inspired by pg_regress in PostgreSQL.

Signed-off-by: Daniel Farina <daniel@fdr.io>