Refinements to make output even more stable
fmtPointer has been completely gutted to prevent accidentally calling
it: doing so will cause a panic.  It was not removed from all paths in
the stable package as of yet, because it is thought that any path that
matters is already treated.

Sort stability on maps has also been fixed by the addition of
retrieving the interface value of a reflect.Value before printing.
The string rendering of reflect.Value contains non-deterministic
information that would be considered in the sort.

Signed-off-by: Daniel Farina <daniel@fdr.io>