Reform parser.go to emit more classical AST
These AST nodes, seen in syntax_types, all have roots in Token values,
which allows for error reporting (including the position of the error)
when the tree is semantically analyzed.

The tests have been updated to reflect these new structures.  Notably
they all now contain positioning information in the emitted structure.

Prior to this, some amount of semantic analysis was done while parsing
(such as converting tokens into integers, et al) in an attempt to end
up with structures more similar to what was predicted to be necessary
at run-time.  This has effectively been removed, and needs to be
re-introduced as a pass on top of the AST, per-normal compiler
architecture.

Signed-off-by: Daniel Farina <daniel@fdr.io>